Cleanup Windows specific projection dev. env.
Remove the non-existent projects and pull in ClusterNode into the
Windows specific development solution for Projections.

Increase projection compile time timeout to 5 seconds

Scope the HandleScope

The HandleScope attempts to track all the local handles allocated within
the handle scope until the handle scope is deleted. In this, we are
deleting resources that it is tracking and once it goes out of scope
it's trying to clean up it's resources which in this case might refer to
the isolate that was already disposed as part of deleting the prelude
script.