Reading registered projections in batches can cause deleted projections to come alive.
Reproduction Steps:

1. Create 200 (double the current batch size) projections and then delete the first projection and don't
specify that you want any of it's state streams deleted.
2. Restart Event Store.
The projection that you have deleted will start up in a stopped state

To remedy this, the way the registration stream is being read has been
changed from the hard to reason about reading the stream backwards
and then using grouping logic to determine whether the projection is
alive/dead.

It has been changed to reading the stream forwards from the beginning and
building up state which represents the projections which are alive.

Another part of this commit addresses the issue of hardcoded strings in
particular those that pertain to the registered projections, namely the
projection registration stream ($projections-$all), the event types such
as $ProjectionCreated, $ProjectionDeleted, $ProjectionsInitialized as
well as the $ProjectionUpdated (which is written to the projection's
configuration stream $projections-{name})