Handle read timeouts during projection recovery
When a projection is in recovery mode, it will verify what is about to
written against what was written.

During the recovery process, the emitted stream will read the stream
backwards and if the read times out for whatever reason, it will cause
the projection to still process events but not write them to the
destination stream as the write is blocked by a flag that indicates that
a read is currently in progress.