Use a unique projection control stream id
Introduce an epoch id into the state change messages that we care about

In the master replication service, when the node is shutting down, call
stop on the TCP Connections as the node is about to shut down anyway.

This was picked up as part of the cluster tests whereby the nodes would
refuse to have an election as the tcp connections were still up.

Use the epoch id as a unique id for the control stream which projections
uses to coordinate when the command readers are ready.

Previously this behaviour was racey as it relied on the command readers
to start up, read the control stream backwards and then read forwards.
The response reader would then start up, write an event to that stream
to which the command readers would reply that they are ready.